# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Welcome to the CMN Express API documentation. This server facilitates seamless integration for managing logistics, tracking, and delivery services efficiently.\nExplore the endpoints to unlock robust features designed to enhance your operational workflow.",
        "title": "CMN Express API Documentation",
        "termsOfService": "https://sharkytech.vercel.app/terms",
        "contact": {
            "name": "Lucas",
            "url": "https://sharkytech.vercel.app",
            "email": "nguyenmanh180102@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "203.145.47.225",
    "basePath": "/",
    "paths": {
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch the details of a user based on their ID. The user can be either a customer or a delivery person.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-information"
                ],
                "summary": "Get user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update for different roles (customer, delivery_person) account_type customer (prepaid, postpaid)\nExample user payload:",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-information"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Request Example",
                        "name": "update_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/express_be_model_req.UpdateUserReq"
                        }
                    }
                ],
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update customer status by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-information"
                ],
                "summary": "Update customer status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Trạng thái mới của khách hàng (accept, deny)",
                        "name": "approval_status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "express_be_model_req.UpdateUserReq": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "account_type": {
                    "description": "2. Field for customer",
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id_card": {
                    "type": "string"
                },
                "identification_number": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longtitude": {
                    "type": "number"
                },
                "middle_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "place_of_origin": {
                    "type": "string"
                },
                "place_of_residence": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "specific_address": {
                    "type": "string"
                },
                "ward": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}